requirements.txt
aniso8601==7.0.0
astroid==2.3.3
autopep8==1.5
Click==7.0
decorator==4.0.11
Flask==1.1.1
Flask-GraphQL==2.0.1
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
imageio==2.1.2
InstagramAPI==1.0.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==0.2.3.2
numpy==1.18.1
Pillow==7.0.0
promise==2.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
requests==2.11.1
requests-toolbelt==0.7.0
Rx==1.6.1
six==1.14.0
tqdm==4.11.2
Werkzeug==0.16.1
wrapt==1.11.2
